+++ FirstGame/Assets/Avoid Obstacle Scripts/Collisonplayer.cs
@@ -13,4 +13,5 @@ public class Collisonplayer : MonoBehaviour
         }
     }
 
+
 }
