+++ FirstGame/Assets/Scripts/AudioManager.cs
@@ -1,0 +1,18 @@
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class AudioManager : MonoBehaviour
+{
+    AudioSource bgmSource;
+    public AudioClip bgmClip;
+
+    void Start()
+    {
+        bgmSource = gameObject.AddComponent<AudioSource>();
+        bgmSource.clip = bgmClip;
+        bgmSource.Play();
+
+    }
+
+
+}
