--- FirstGame/Assets/Scripts/BackgroundMusic.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 39c37edace15b99449246bedd2f273e2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
