+++ FirstGame/Assets/Sound/Wonderball - 인게임 기본 (신남, 발랄함, 엉뚱함).mp3.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: db5d11668ffd75644bc7d0fd6dad5165
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 7
+  defaultSettings:
+    serializedVersion: 2
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+    preloadAudioData: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
